// SeatLayoutND/src/app/models/printticket.ts
  
export class PrintTicket{
    TicketID: number;
    BusID: number;
    BookingUserID: number;
    FromLocation: string;
    ToLocation: string;
    FromDate: string;
    ToDate: string;
    FromTime: string;
    ToTime: string;
    TicketBookingStatus: number;
    NumberOfSeats: number;
    Fare: number;
    PaymentStatus: number;
}